// nextjs/types.ts
export interface User {
  id: number;
  name?: string; // '?' indicates that the field is optional
  email: string;
  emailVerified?: string;
  image?: string;
  Password?: string;
  createdAt: string;
  updatedAt: string;
  favoriteIds: string[];
  phone_number: string;
  User_type: string;
  orders: Order[];
  accounts: Account[];
  sessions: Session[];
  Post: Post[];
  Comment: Comment[];
  streetAddress: string;
}

export interface Account {
  id: number;
  userId: string;
  type: string;
  provider: string;
  providerAccountId: string;
  refresh_token?: string;
  access_token?: string;
  expires_at?: number;
  token_type?: string;
  scope?: string;
  id_token?: string;
  session_state?: string;

  user: User;
}
// ...
export interface Session {
  id: string;
  sessionToken: string;
  userId: string;
  expires: string;
  user: User;
}

export interface UserBlog {
  id: string;
  name?: string | null;
  email: string;
  emailVerified?: string | null;
  image?: string | null;
  accounts: Account[];
  sessions: Session[];
  Post: Post[];
  Comment: Comment[];
}

export interface VerificationToken {
  identifier: string;
  token: string;
  expires: string;
}

export interface CategoryBlog {
  id: string;
  slug: string;
  title: string;
  img?: string | null;
  Posts: Post[];
}

export interface Post {
  id: string;
  createdAt: string;
  slug: string;
  title: string;
  desc: string;
  img?: string | null;
  views: number;
  catSlug: string;
  cat: CategoryBlog;
  userEmail: string;
  user: UserBlog;
  comments: Comment[];
}

export interface Comment {
  id: string;
  createdAt: string;
  desc: string;
  userEmail: string;
  user: User;
  postSlug: string;
  post: Post;
}